Update puppet-ssh for client config managment
Update puppet-ssh to support configuring and managing the
/etc/ssh/ssh_config file with optimisations for openstack.org
ephemeral hosts (do not store the host keys in the known_hosts
files).

As part of this work, support of the SSH server configuration
management should occur in the "ssh::server" module; puppet-ssh now
supports either using "ssh" or "ssh::server" for the transitional
period.

Future changes will look at broadening the configuration options
and validation of the configuration options for the client config
such as richer host matching, stricthostkeychecking with pre-defined
keys distributed, etc.

Change-Id: Ic2dcc00945ab1eed985eb32dcfae910f4363cacb